pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "thesis"
version = "0.1.0"
description = "Deterministic AV ethics thesis package with scenario parsing and mathematical risk analysis."
requires-python = ">=3.10"
dependencies = [
  "gradio>=5.0",
  "langchain>=1.2",
  "langchain-chroma>=1.1",
  "langchain-community>=0.4",
  "langchain-openai>=1.1",
  "langchain-text-splitters>=1.1",
  "python-dotenv>=1.1",
  "pypdf>=6.7",
  "tiktoken>=0.9",
]

[dependency-groups]
dev = [
  "ipykernel>=6.29",
  "jupyter>=1.1",
]

[tool.setuptools.packages.find]
include = ["thesis*"]
